[TEST] Split integration tests related to DOM checks
Create dedicated files for each kind of tests for clarity.

Configure mxGraph label generation during jest setup. This ensures that all
tests have the right configuration (no more setup in test files).
