+++ test/integration/config/mxgraph-config.ts
@@ -1,0 +1,21 @@
+/**
+ * Copyright 2021 Bonitasoft S.A.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { mxgraph } from '../../../src/component/mxgraph/initializer';
+
+// Force usage of ForeignObject
+// By default, mxGraph detects no ForeignObject support when running tests in jsdom environment
+mxgraph.mxClient.NO_FO = false;
